+++ src/scripts/parts/todolist.js
@@ -1,5 +1,12 @@
 import $ from "jquery";
 import {createItem} from "./todolist/view";
+import {
+    storeNote,
+    getTodolist,
+    deleteItemById,
+    setFinishTodolist,
+    editItemById
+} from "./todolist/model";
 
 function init() {
     console.log("Hello World! Todolist!");
@@ -15,23 +22,8 @@ function init() {
         deleteItem($(this).attr("value"));
     });
     $("ul").on("click", ".check-btn", function () {
-        console.log("check");
         checkItem($(this).attr("value"));
     });
-}
-
-//api return all todo list in an array
-function getTodolist() {
-    var Todolist = [];
-    Todolist = JSON.parse(localStorage.getItem("noteStr"));
-
-    return Todolist;
-}
-//api give an id and return event status
-function finishTodolist(id) {
-    var TodolistJSON;
-    TodolistJSON = JSON.parse(localStorage.getItem(id));
-    return TodolistJSON.status;
 }
 
 function initPage() {
@@ -50,41 +42,18 @@ function initPage() {
 }
 
 function addNote() {
+    // 取新增的資料
     var inputValue = $("#add-text").val();
-    if (inputValue != "") {
-        storeNote(inputValue);
-        $("#add-text").val("");
+    $("#add-text").val("");
+    // 防呆
+    if (inputValue == "") {
+        return ;
     }
+    var params = storeNote(inputValue);
+    $("ul").append(createItem(params.id, params.text, params.status));
 }
-function storeNote(text) {
-    var noteArray = [];
-    var createTime = getCurrentTime();
-    var id = 1;
-    if (localStorage.getItem("noteStr") != null) {
-        noteArray = JSON.parse(localStorage.getItem("noteStr"));
-        id = noteArray[noteArray.length - 1].id + 1;
-    }
-    var noteJson = {
-        id: id,
-        name: text,
-        status: false,
-        createAt: createTime,
-        finishAt: null
-    };
-    noteArray.push(noteJson);
-    localStorage.setItem("noteStr", JSON.stringify(noteArray));
-    $("ul").append(createItem(id, text, false));
-}
-function getCurrentTime() {
-    var today = new Date();
-    var date =
-        today.getFullYear() + "-" + (today.getMonth() + 1) + "-" + today.getDate();
-    var time =
-        today.getHours() + ":" + today.getMinutes() + ":" + today.getSeconds();
-    var dateTime = date + " " + time;
 
-    return dateTime;
-}
+
 function showNotFinishList() {
     $(".top-left .left").css("opacity", "0.5");
     $(".top-left .right").css("opacity", "1");
@@ -112,19 +81,8 @@ function checkItem(id) {
     }
 }
 
-function setFinishTodolist(id) {
-    var noteArray = getTodolist();
-    for (var i = 0; i < noteArray.length; i++) {
-        if (noteArray[i].id == id) {
-            noteArray[i].status = true;
-            localStorage.setItem("noteStr", JSON.stringify(noteArray));
-            return true;
-        }
-    }
-    return false;
-}
+
 function editItem(id) {
-    var noteArray = getTodolist();
     Swal.fire({
         title: "請輸入修改記事",
         input: "text",
@@ -135,35 +93,17 @@ function editItem(id) {
             }
         }
     }).then(result => {
-        if (result.value) {
-            //console.log(result.value);
-            for (var i = 0; i < noteArray.length; i++) {
-                if (noteArray[i].id == id) {
-                    if (newText) {
-                        noteArray[i].name = result.value;
-                        localStorage.setItem("noteStr", JSON.stringify(noteArray));
-                        break;
-                    }
-                }
-            }
-            initPage();
+        if (!result.value) {
+            return ;
         }
+        editItemById(id, result.value);
+        initPage();
     });
 }
+
 function deleteItem(id) {
     $("li." + id).remove();
-    var noteArray = [];
-    noteArray = getTodolist();
-    for (var i = 0; i < noteArray.length; i++) {
-        if (id == noteArray[i].id) {
-            noteArray.splice(i, 1);
-            localStorage.setItem("noteStr", JSON.stringify(noteArray));
-            if (noteArray.length == 0) {
-                localStorage.removeItem("noteStr");
-            }
-            break;
-        }
-    }
+    deleteItemById(id);
 }
 
-export { init, finishTodolist, setFinishTodolist };
+export { init, setFinishTodolist };
